:art: style(entity,dto,service,...): 实体等pojo主键变为Long类型

+++ src/main/java/com/rymcu/forest/lucene/service/UserLuceneService.java
@@ -70,5 +70,5 @@ public interface UserLuceneService {
    * @param ids 用户id(半角逗号分隔)
    * @return
    */
-  List<UserDTO> getUsersByIds(Integer[] ids);
+  List<UserDTO> getUsersByIds(Long[] ids);
 }
